api/requirements.txt
arrow
fastapi
gunicorn
loguru
notifiers
pydantic
pyotp
python-dotenv
uvicorn
